Minor changes
Changed some functions I had made to protected instead of public as it
shouldn't be used outside of the package (mainly the set functions, for
actually updating profile information, you would need to use the
updateProfile function). Remade search function to return an ArrayList
of user instead of a String[][] array as requested. Tested it and it
works. Removed the Businesses class as it was a hold over from the old
DB method that never got removed. Commented out the SessionTester in
TestMain for now.